% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alias_cpp}
\alias{alias_cpp}
\title{Call alias_c function from C}
\usage{
alias_cpp(x, n_draws)
}
\arguments{
\item{x}{A vector of unnormalized probabilities}

\item{n_draws}{Integer number of draws from the discrete distribution}
}
\value{
An integer-valued vector of draws from the discrete distribution
}
\description{
Call alias_c function from C
}
