% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_test.R
\name{alias_R}
\alias{alias_R}
\title{Test alias C code for discrete random numbers}
\usage{
alias_R(p, n_draws)
}
\arguments{
\item{p}{Vector of unnormalised probabilities}

\item{n_draws}{Number of random draws to make}
}
\value{
d Vector of random draws
}
\description{
Test alias C code for discrete random numbers
}
