% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{eff}
\alias{eff}
\title{Compute efficiency information for a skew draw evaluation}
\usage{
eff(lnf, true_lnf, z_max, n_pos)
}
\arguments{
\item{lnf, }{log normalized density evaluated on a grid}

\item{true_lnf, }{true log (but unnormalized) density evaluated on the same grid}

\item{z_max, }{maximum grid point}

\item{n_pos, }{number of positive grid points (total number is 2*n_pos + 1)}
}
\value{
A list of outputs
}
\description{
\code{eff} does ...
}
