% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Po_GaPo.R
\name{Po_GaPo}
\alias{Po_GaPo}
\title{Compute info for Poisson and Gamma-Poisson models}
\usage{
Po_GaPo(n, y_bar, r, theta, omega, mode = 0, x_max, n_pos = 1000)
}
\arguments{
\item{n, }{number of observations}

\item{y_bar, }{mean count}

\item{r, }{Gamma-Poisson parameter, r=inf is Poisson special case}

\item{theta, }{multiplicative scale parameter for Poisson or Gamma-Poisson}

\item{omega, }{prior precision}

\item{mode, }{mode of distribution}

\item{x_max, }{maximum grid-point value}

\item{n_pos, }{number of positive grid points}
}
\value{
A list of outputs
}
\description{
\code{Po_GoPo} does ...
}
