% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skew_draw_cpp}
\alias{skew_draw_cpp}
\title{Call skew_draw_c function in skew.c}
\usage{
skew_draw_cpp(n_grid_points, code, mode, h, mu, omega, n_draws)
}
\arguments{
\item{n_grid_points, }{number of spline grid points in f_u approximation}

\item{code, }{0 for spline_skew_draw without v transformation, 1 for spline_skew_draw with v transformation,
2 for old skew_draw}

\item{mode}{A scalar, the mode of the target distribution}

\item{h}{A vector, the first five derivatives of the log target distribution}

\item{mu}{A scalar, the prior mean}

\item{omega}{A scalar, the prior precision}

\item{n_draws}{A scalar integer, the number of draws to make}
}
\value{
A list consisting of a vector of draws from the target distribution and a vector of corresponding
log density evaluations.
}
\description{
Call skew_draw_c function in skew.c
}
