% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spline_eval_cpp}
\alias{spline_eval_cpp}
\title{Call C language spline_eval_c function}
\usage{
spline_eval_cpp(p, m, u)
}
\arguments{
\item{p}{A vector of levels on grid}

\item{m}{A vector of derivatives on grid}

\item{u}{A vector of values at which to evaluate a spline density}
}
\value{
The vector of spline density values
}
\description{
Call C language spline_eval_c function
}
