% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skew_draw_cpp}
\alias{skew_draw_cpp}
\title{Call skew_draw_c function in skew.c}
\usage{
skew_draw_cpp(mode, h, mu, omega, n_draws)
}
\arguments{
\item{mode}{A scalar, the mode of the target distribution}

\item{h}{A vector, the first five derivatives of the log target distribution}

\item{mu}{A scalar, the prior mean}

\item{omega}{A scalar, the prior precision}

\item{n_draws}{A scalar integer, the number of draws to make}
}
\value{
A vector of draws from the target distribution
}
\description{
Call skew_draw_c function in skew.c
}
